# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracer77 API",
        "description": "API d'authentification et de géolocalisation pour Tracer77",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.77:8000/",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/family/create": {
            "post": {
                "summary": "Créer un espace famille",
                "operationId": "ea3c1b20b11ed4ad760837f264931089",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Famille Dupont"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Famille créée"
                    },
                    "409": {
                        "description": "Appartient déjà à une famille"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/family/invite": {
            "post": {
                "summary": "Inviter un membre dans l'espace famille",
                "operationId": "d88e4622fd487e91f5b28518bcdb90b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "membre-test@exemple.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation envoyée"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/family/accept-invite": {
            "post": {
                "summary": "Accepter une invitation à rejoindre une famille",
                "operationId": "84f18b08d70a49ba99f3955f4da75a51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "le_token_recu_par_email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitation acceptée"
                    },
                    "404": {
                        "description": "Invitation invalide"
                    },
                    "410": {
                        "description": "Invitation expirée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/family/members": {
            "get": {
                "summary": "Lister les membres et enfants de la famille",
                "operationId": "35ea2781ff2c6cbcf7cce7c7ca2806e8",
                "responses": {
                    "200": {
                        "description": "Liste retournée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/family/members/{id}": {
            "delete": {
                "summary": "Retirer un membre de la famille",
                "operationId": "e2094d61d52052501af0800d787b293b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Membre retiré"
                    },
                    "403": {
                        "description": "Non autorisé"
                    },
                    "404": {
                        "description": "Membre introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/toggle-position-sharing": {
            "post": {
                "summary": "Activer/désactiver le partage de sa position",
                "operationId": "40d6d003810978f478766a4856bb94c9",
                "responses": {
                    "200": {
                        "description": "Préférence mise à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum ici",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}